feat: Add ability to disable automatic bookkeeping
Added a new config variable, autoUpdateEnabled, for the config object of
the ChangeDetector constructor. I defaulted it to true, to make the
class behave the same as was prior to the update by default.

When disabled, the onus is now on the consumer of the API to invoke
updateRecords themselves with the records that they want to update the
bookkeeping on. This is useful as it allows for repeatability upon
failures.

I updated the README with details of the new option.

I also bumped the semantic version in package.json from 1.1.0 to 1.2.0
as I believe this is a new feature, but doesn't break existing API
usage.

This commit will resolve #2